pyproject.toml
[project]
name = "financebud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.0",
    "alembic>=1.12.0",
    "black>=23.0.0",
    "fastapi>=0.104.0",
    "fastmcp>=0.9.0",
    "httpx>=0.25.0",
    "httpx-sse>=0.4.0",
    "isort>=5.12.0",
    "jinja2>=3.1.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "pandas>=1.5.0",
    "pydantic>=2.4.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "wsproto>=1.2.0",
]
